requirements.txt
typer
PyYAML
httpx
tenacity
feedparser
beautifulsoup4
trafilatura
readability-lxml
crawl4ai
ollama
psycopg
pymongo
pytest
